Move protobuf spec to lib/
Tonic seems to copy the protobuf definition around at compile time,
causing cargo to think the cache is invalid.

Therefore this commit moves the protobuf file to lib/ when tonic would
end up copying it too stopping cargo from incorrectly recomciling.

Extra:

Move custom resources to lib/
