+++ controller/Cargo.toml
@@ -28,11 +28,6 @@ doc = false
 name = "data_set_crdgen"
 path = "src/data_set_crd_gen.rs"
 
-[lib]
-name = "controller"
-path = "src/lib.rs"
-
-
 [dependencies]
 ame = { version = "0.1.0", path = "../lib", features = ["ame-control", "custom-resources"] }
 async-trait = "0.1.66"
@@ -40,7 +35,7 @@ duration-string = "0.2.0"
 either = "1.8.0"
 envconfig = "0.10.0"
 futures = "0.3.25"
-git2 = "0.15.0"
+git2 = "0.17.0"
 humantime = "2.1.0"
 k8s-openapi = { version = "0.17.0", features = ["v1_23", "schemars" ] }
 kube = { version = "0.80.0", features = ["runtime", "client", "derive"] }
@@ -48,11 +43,7 @@ reqwest = { version = "0.11.14", features = ["json"] }
 schemars = "0.8.11"
 serde = {version = "1.0.147", features = ["derive"]}
 serde_json = "1.0.87"
-serde_merge = "0.1.3"
-serde_tuple = "0.5.0"
 serde_yaml = "0.9.14"
-similar = "2.2.1"
-similar-asserts = "1.4.2"
 thiserror = "1.0.37"
 tokio = { version = "1.21.2", features = ["macros", "rt-multi-thread"] }
 tonic = "0.8.3"
@@ -60,8 +51,4 @@ tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.16", features = ["env-filter"] }
 
 [dev-dependencies]
-assert_fs = "1.0.10"
-insta = {version = "1.21.0", features = ["yaml"] }
-predicates = "2.1.4"
-serial_test = "0.9.0"
 k8s-openapi = { version = "0.17.0", features = ["v1_23", "schemars" ] }
