+++ controller/src/crdgen.rs
@@ -1,8 +1,6 @@
+use ame::custom_resources::task::Task;
 use kube::CustomResourceExt;
 
 fn main() {
-    print!(
-        "{}",
-        serde_yaml::to_string(&controller::Task::crd()).unwrap()
-    );
+    print!("{}", serde_yaml::to_string(&Task::crd()).unwrap());
 }
