+++ controller/src/project_crdgen.rs
@@ -1,4 +1,4 @@
-use controller::project::Project;
+use ame::custom_resources::project::Project;
 use kube::CustomResourceExt;
 fn main() {
     print!("{}", serde_yaml::to_string(&Project::crd()).unwrap())
