+++ service/Cargo.toml
@@ -16,7 +16,6 @@ tokio = { version = "1.0", features = ["rt-multi-thread", "macros", "sync", "tim
 tokio-stream = "0.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-controller = { path = "../controller" }
 k8s-openapi = { version = "0.17.0", features = ["v1_23", "schemars"] }
 kube = { version = "0.80.0", features = ["runtime", "client" ] }
 thiserror = "1.0.37"
@@ -41,5 +40,4 @@ prost-build = "0.11.2"
 serde = "1.0"
 
 [dev-dependencies]
-controller = { path = "../controller" }
 serial_test = "0.9.0"
