+++ shell.nix
@@ -23,6 +23,7 @@ pkgs.mkShell {
     grpcurl
     sass
     rust-analyzer
+    jq
   ];
 
   # Certain Rust tools won't work without this
